+++ packages/alfa-lazy/src/lazy.ts
@@ -30,11 +30,19 @@ export class Lazy<T> implements Monad<T>, Functor<T>, Equatable, Serializable {
   }
 
   public map<U>(mapper: Mapper<T, U>): Lazy<U> {
-    return new Lazy(this._value.map(mapper));
+    return this.flatMap((value) => Lazy.force(mapper(value)));
   }
 
   public flatMap<U>(mapper: Mapper<T, Lazy<U>>): Lazy<U> {
-    return new Lazy(this._value.flatMap((value) => mapper(value)._value));
+    return new Lazy(
+      this._value.flatMap((value) => {
+        if (this._value.isSuspended()) {
+          this._value = Trampoline.done(value);
+        }
+
+        return mapper(value)._value;
+      })
+    );
   }
 
   public equals(value: unknown): value is this {